requirements.txt
# Core ML & Data
pandas>=2.1.0
numpy>=1.24.0
lightgbm>=4.1.0
xgboost>=2.0.0
scikit-learn>=1.3.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0  # For JS-heavy pages if needed

# Data Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver

# API & Monitoring
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Geospatial
geopy>=2.4.0
folium>=0.15.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0  # Faster fuzzy matching

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Jupyter (for exploration)
jupyter>=1.0.0
ipykernel>=6.26.0

# Streamlit UI (separate from core)
streamlit>=1.28.0
streamlit-folium>=0.15.0
plotly>=5.18.0